Updated utf_length_n() to count NUL bytes as normal characters.
Previously utf_length_n() would stop at the first NUL byte, even if ‘len’
said that more data was available.  This conforms with the semantics of,
for example, strnlen(), but as we in Ruby can easily get Strings that
contain any byte, we can’t treat NUL bytes separately in the case where
‘len’ is provided (as opposed to utf_length() that is).  So all *_n()
functions will have to be rewritten to handle NUL bytes like any other
byte.  We begin with the utf_length_n() function, because it is an obvious
case where this was causing a problem.