Fix String#upcase in lithuanian locale.
The problem was that accents were being doubled for ‘i’s in the lithuanian
locale.  The reason for this was that the pointer into the string being
upcased wasn’t being advanced properly when output the correct combining
characters, so previous combining characters (from the lowercase version)
would linger.  The fix was to make sure that the pointer into the string
was advanced properly.

This problem was spotted by Lugovoi Nikolai.